BOCM-486 - moves blob store connector to own file
