+++ apps/document-storage/src/server/app/controller.js
@@ -1,23 +1,11 @@
-import getStream from 'into-stream';
-import md5 from 'crypto-js/md5.js';
-import { BlobServiceClient } from '@azure/storage-blob';
-import config from '../config/config.js';
-
-const connectionString = config.blobStore.connectionString;
-const containerName = config.blobStore.container;
-
-const getBlobName = originalName => {
-    const identifier = Math.random().toString().replace(/0\./, '');
-    return `${identifier}-${originalName}`;
-};
+import * as blobStoreService from './blob-store/service.js';
 
 export const getAllDocuments = async function(req, res, next) {
-    const blobServiceClient = BlobServiceClient.fromConnectionString(connectionString);
-    const containerClient = blobServiceClient.getContainerClient(containerName);
-    const listBlobsResponse = await containerClient.listBlobFlatSegment();
+    
+    const blobsResponse = await blobStoreService.getListOfBlobs();
 
     const blobs = [];
-    for await (const blob of listBlobsResponse.segment.blobItems) {
+    for await (const blob of blobsResponse.segment.blobItems) {
         blobs.push({ name: blob.name, metadata: blob.metadata });
     }
 
@@ -25,21 +13,7 @@ export const getAllDocuments = async function(req, res, next) {
 }
 
 export const uploadDocument = async function(req, res) {
-    const blobServiceClient = BlobServiceClient.fromConnectionString(connectionString);
-    const blobName = getBlobName(req.file.originalname);
-    const stream = getStream(req.file.buffer);
-
-    const md5Value = Uint8Array.from(md5(stream).toString());
-
-    const containerClient = blobServiceClient.getContainerClient(containerName);;
-    const blockBlobClient = containerClient.getBlockBlobClient(blobName);
-    await blockBlobClient.uploadStream(stream,
-        undefined, undefined,
-        {
-            blobHTTPHeaders: {
-                blobContentType: "application/json",
-                blobContentMD5: md5Value
-            }
-        });
+    await blobStoreService.uploadBlob(req.file.originalname, req.file.buffer);
+    
     res.send({ message: 'File uploaded to Azure Blob storage.' });
 }
